Potential fix for code scanning alert no. 99: Multiplication result converted to larger type

+++ pc-bios/s390-ccw/virtio-blkdev.c
@@ -98,7 +98,7 @@ unsigned long virtio_load_direct(unsigned long rec_list1, unsigned long rec_list
     if (status) {
         return 0;
     }
-    addr += sec_num * virtio_get_block_size();
+    addr += (unsigned long)sec_num * virtio_get_block_size();
 
     return addr;
 }
